fixup! + ProductsTableCellDelegate
